requirements.txt
streamlit
pandas
requests
plotly
pyyaml
sqlite3-binary
